Allow to configure text partitioning values from appsettings.json Move KMB extensions to Abstractions
