+++ nuget-package.props
@@ -1,7 +1,7 @@
 ﻿<Project>
     <PropertyGroup>
         <!-- Central version prefix - applies to all nuget packages. -->
-        <Version>0.25.0</Version>
+        <Version>0.26.0</Version>
 
         <!-- These are set at the project level-->
         <IsPackable>false</IsPackable>
