+++ service/Service/appsettings.json
@@ -63,6 +63,15 @@
       ],
       // "None" or "AzureAIDocIntel"
       "ImageOcrType": "None",
+      // Partitioning / Chunking settings
+      "TextPartitioning": {
+        // Maximum size of paragraphs, in tokens. Tokens depend on the LLM in use.
+        "MaxTokensPerParagraph": 1000,
+        // Maximum size of lines of text, in tokens. Tokens depend on the LLM in use.
+        "MaxTokensPerLine": 300,
+        // How many tokens from a paragraph to keep in the following paragraph.
+        "OverlappingTokens": 100
+      },
       // Note: keep the list empty in this file, to avoid unexpected merges
       // with the list defined in appsettings.*.json.
       // If the list is empty, KernelMemoryConfig uses 'Constants.DefaultPipeline'.
